backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
elevenlabs
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pydantic-settings
python-dotenv
python-multipart
